Move SplitTooLongLine to its own module
This commit aligns SplitTooLongLine with
a0d2b4bf67cf331b66e24d509770eeeccca6be78.
